% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hrate.R
\name{calculate_hrate}
\alias{calculate_hrate}
\title{Calculates back-exchange based 
on the Hamuro table}
\usage{
calculate_hrate(sequence)
}
\description{
Calculates back-exchange based 
on the Hamuro table
}
\examples{
calculate_hrate("GFGDLKSPAGL")

}
