% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subsections.R
\name{create_subsections}
\alias{create_subsections}
\title{Create subsections from peptides}
\usage{
create_subsections(dat)
}
\arguments{
\item{dat}{peptides data}
}
\description{
Create subsections from peptides
}
\examples{
dat <- filter(alpha_dat, State == "Alpha_KSCN", End < 35)
subsections <- create_subsections(dat)

}
