% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subsections_dataset.R
\name{get_subsection_data}
\alias{get_subsection_data}
\title{calculate mass uptake for one subsection}
\usage{
get_subsection_data(dat, state = dat[["State"]][1], subsection)
}
\description{
calculate mass uptake for one subsection
}
\examples{
subsections <- create_subsections(alpha_dat)
subsection <- subsections[3, ]
get_subsection_data(dat = alpha_dat, subsection = subsection)

}
