% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uc_scaled.R
\name{plot_uc_scaled}
\alias{plot_uc_scaled}
\title{Plots UC scaled with retention}
\usage{
plot_uc_scaled(kin_dat, state = unique(kin_dat[["State"]])[1], ret = 1)
}
\arguments{
\item{kin_dat}{uptake dat for a peptide}

\item{ret}{scaling parameter

This function plots two deuterium uptake curve:
experimental and scaled by the retention parameter}
}
\description{
Plots UC scaled with retention
}
