% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uc_with_origin.R
\name{plot_uc_with_origin}
\alias{plot_uc_with_origin}
\title{Plot uc of subfragment with parent peptides}
\usage{
plot_uc_with_origin(dat, subsection_dat, subsection, state = dat[["State"]][1])
}
\arguments{
\item{dat}{original dat}

\item{subsection}{record of chosen subsection}
}
\description{
Plot uc of subfragment with parent peptides
}
\examples{
subsections <- create_subsections(alpha_dat)
subsection <- subsections[3,]
subsection_dat <- get_subsection_data(alpha_dat, subsection = subsection)

plot_uc_with_origin(dat = alpha_dat, subsection_dat = subsection_dat)

}
