% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subsections_dataset.R
\name{create_subsections_dataset}
\alias{create_subsections_dataset}
\title{to test}
\usage{
create_subsections_dataset(dat, subsections)
}
\arguments{
\item{dat}{raw data}

\item{subsections}{subsections created from 
the dat coverage}
}
\description{
use get_subsection_data
}
\examples{
dat <- filter(alpha_dat, State == "Alpha_KSCN", End < 35)
subsections <- create_subsections(dat)
subs_dats <- create_subsections_dataset(dat, subsections)

}
